Change items file to be a CSV file
As discussed somewhere, spaces in paths makes the original format for items file not work. So, changing it to a csv file with the hope/assumption that folks are less likely to want to put commas into their file names than they are to put spaces into their file names.