Replace usage of Fair Research Login
This completely gets rid of Fair Research Login and uses the Globus
SDK native client directly.

It also puts a new token storage mechanism in place keyed entirely by
scope strings. The intent is that this should allow for management of
dynamic dependency tokens better as we'll treat the entire scope
string as the key including the dynamic dependency portion. This may
be non-optimal as we may have more tokens than we actually need, but
it should make sure that on every operation we get a token that is
sufficient for the full dynamic dependency string requested.