Make the job transfer example runnable
If run with most unix shells, the --item paths will have the local
user's home directory interpolated for ~. This will almost always
result a transfer task that fails. Also, the source files must be
populated by the user before running the example.

This patch slightly updates the text before the example, and also
makes the example runnable by quoting the ~ strings and using the
/share/godata directory for the source.